fix(backend): corregir problemas al arrancar el servidor

+++ backend/src/modules/users/controllers/admin.controller.ts
@@ -2,12 +2,16 @@ import { RequestHandler } from 'express';
 import { responseBuilder } from '../../../core/common/response-builder';
 import { AdminService } from '../services/admin.service';
 
+type AdminParams = {
+  id: string;
+};
+
 export class AdminController {
   constructor(
     private adminService: AdminService,
   ) { }
 
-  deleteById: RequestHandler = async (req, res, next) => {
+  deleteById: RequestHandler<AdminParams> = async (req, res, next) => {
     try {
       const id = req.params.id;
       await this.adminService.delete(id);
